@@ -18,55 +18,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.1",
-    "@angular/common": "6.0.1",
-    "@angular/compiler": "6.0.1",
-    "@angular/core": "6.0.1",
-    "@angular/forms": "6.0.1",
-    "@angular/http": "6.0.1",
-    "@angular/platform-browser": "6.0.1",
-    "@angular/platform-browser-dynamic": "6.0.1",
-    "@angular/platform-server": "6.0.1",
-    "@angular/router": "6.0.1",
-    "adal-angular": "1.0.17",
-    "applicationinsights-js": "1.0.8",
+    "@angular/animations": "~8.2.11",
+    "@angular/cdk": "~8.2.3",
+    "@angular/common": "~8.2.11",
+    "@angular/compiler": "~8.2.11",
+    "@angular/core": "~8.2.11",
+    "@angular/forms": "~8.2.11",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.11",
+    "@angular/platform-browser-dynamic": "~8.2.11",
+    "@angular/router": "~8.2.11",
+    "hammerjs": "^2.0.8",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1",
+    "adal-angular4": "~4.0.12",
+    "applicationinsights-js": "~1.0.20",
     "bootstrap": "3.3.7",
     "core-js": "2.5.4",
-    "expose-loader": "0.7.5",
     "font-awesome": "4.7.0",
     "less": "3.0.4",
-    "lodash": "4.17.10",
+    "lodash": "~4.17.15",
     "moment": "2.22.1",
-    "primeng": "4.3",
-    "rxjs": "^6.1.0",
-    "rxjs-compat": "^6.1.0",
-    "zone.js": "^0.8.26"
+    "primeicons": "1.0.0",
+    "primeng": "~7.0.0"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "6.0.1",
-    "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "~6.0.1",
-    "@angular/language-service": "^6.0.0",
-    "@types/jasmine": "2.8.7",
+    "@angular-devkit/build-angular": "~0.803.14",
+    "@angular/cli": "~8.3.14",
+    "@angular/compiler-cli": "~8.2.11",
+    "@angular/language-service": "~8.2.11",
     "@types/node": "~8.9.4",
-    "@types/adal": "^1.0.29",
-    "@types/adal-angular": "1.0.1",
-    "@types/applicationinsights-js": "1.0.2",
-    "@types/lodash": "^4.14.108",
-    "typescript": "2.7.2",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "3.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.2",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "1.1.0",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-spec-reporter": "0.0.32",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1"
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/adal-angular": "~1.0.1",
+    "@types/applicationinsights-js": "~1.0.9",
+    "@types/lodash": "~4.14.144",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-junit-reporter": "~2.0.1",
+    "karma-spec-reporter": "~0.0.32",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
