@@ -1,11 +1,11 @@
 import { TestInjector } from '../../demo-common/testing/testing-helpers';
 import { ActivatedRoute } from '@angular/router';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async } from '@angular/core/testing';
 import { PageErrorComponent } from './page-error.component';
-import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs';
 
 class MockActivatedRoute extends ActivatedRoute {
-  queryParams = Observable.of({
+  queryParams = of({
     errorMessage: 'Test message'
   });
 }
